backend/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
annotated-doc==0.0.4
annotated-types==0.6.0
anthropic==0.86.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
bidict==0.23.1
blinker==1.6.2
blis==0.7.11
branca==0.6.0
cachelib==0.12.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.1.0
cryptography==46.0.6
cycler==0.11.0
cymem==2.0.8
distro==1.9.0
docstring_parser==0.17.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.135.2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
folium==0.14.0
fonttools==4.42.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.7.1
httpx==0.27.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.13.0
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.0
MarkupSafe==2.1.3
matplotlib==3.7.2
msgspec==0.18.6
multidict==6.0.5
murmurhash==1.0.10
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
pandas==2.0.3
patsy==0.5.3
pefile==2023.2.7
Pillow==10.0.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.5.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyJWT==2.8.0
PyMySQL==1.1.2
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.2.2
python-engineio==4.9.1
python-multipart==0.0.22
python-socketio==5.11.2
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.5
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==1.0.0
statsmodels==0.14.0
thinc==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
twilio==9.2.3
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.42.0
wasabi==1.1.2
watchfiles==1.1.1
weasel==0.3.4
websockets==16.0
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.4
